Add C++ solution for Inverse_Inversions

+++ Construction_Problems/Inverse_Inversions/Inverse_Inversions.cpp
@@ -8,58 +8,59 @@
 #include<bits/stdc++.h>
 using namespace std;
  
-void rotate(vector<int>&a, int num) {
-	int curr = num + 1;
-	int n = a.size();
-	for(int i = 0; i < (n - num); i++) {
+void rotate(vector<long long>&a, long long num) {
+	long long curr = num + 1;
+	long long n = a.size();
+	for(long long i = 0; i < (n - num); i++) {
 		a[i] = curr;
 		curr++;
 	}
 	a[n - num] = num;
-	for(int i = n - num + 1; i < n; i++) {
+	for(long long i = n - num + 1; i < n; i++) {
 		a[i] = a[i - 1] - 1;
 	}
 }
  
-void rotate2(vector<int>&a, int rem, int num) {
-	for(int i = 0; i < rem; i++) {
+void rotate2(vector<long long>&a, long long rem, long long num) {
+	for(long long i = 0; i < rem; i++) {
 		a[i] = a[i + 1];
 	}
 	a[rem] = num;
 }
  
 int main() {
-	int n, k;
+	long long n, k;
 	cin >> n >> k;
-	vector<int>a(n);
-	for(int i = 0; i < n; i++) {
+	vector<long long>a(n);
+	for(long long i = 0; i < n; i++) {
 		a[i] = (i + 1);
 	}
-    if(k == (n * (n - 1)) / 2){
-        for(int i = n; i >= 1; i--){
+ 
+    if(k == ((n * (n - 1)) / 2)){
+        for(long long i = n; i >= 1; i--){
             cout << i << " ";
         }
         cout << "\n";
         return 0;
     }
-	for(int x = 1; x <= ((n - 1) * n) / 2; x++) {
+	for(long long x = 1; x <= ((n - 1) * n) / 2; x++) {
 		if(x > k) {
 			break;
 		}
 		else if(x == k) {
-			int div = (x + n - 2)/ (n - 1);
-			int rem = x % (n - 1);
+			long long div = (x + n - 2)/ (n - 1);
+			long long rem = x % (n - 1);
             if(rem == 0){
                 rem = (n - 1);
             }
-			for(int num = 1; num < div; num++) {
+			for(long long num = 1; num < div; num++) {
 				rotate(a, num);
 			}
  
 			rotate2(a, rem, div);
 		}
 	}
-	for(int &e : a) {
+	for(long long &e : a) {
 		cout << e << " ";
 	}
